pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "leaf-measure"
version = "0.1.0"
description = "Portable FAMeLeS-based leaf morphology measurement for agent workflows."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
classifiers = [
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "Pillow>=10.0",
  "scipy>=1.13",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
packages = ["engine"]
